Fixed manifest imports
