+++ nemo/collections/asr/models/clustering_diarizer.py
@@ -29,10 +29,6 @@ from tqdm import tqdm
 from nemo.collections.asr.models.classification_models import EncDecClassificationModel
 from nemo.collections.asr.models.label_models import EncDecSpeakerLabelModel
 from nemo.collections.asr.parts.mixins.mixins import DiarizationMixin
-from nemo.collections.asr.parts.utils.manifest_utils import (
-    segments_manifest_to_subsegments_manifest,
-    write_rttm2manifest,
-)
 from nemo.collections.asr.parts.utils.speaker_utils import (
     audio_rttm_map,
     get_embs_and_timestamps,
@@ -40,7 +36,9 @@ from nemo.collections.asr.parts.utils.speaker_utils import (
     parse_scale_configs,
     perform_clustering,
     score_labels,
+    segments_manifest_to_subsegments_manifest,
     validate_vad_manifest,
+    write_rttm2manifest,
 )
 from nemo.collections.asr.parts.utils.vad_utils import (
     generate_overlap_vad_seq,
