+++ nemo/collections/asr/parts/utils/diarization_utils.py
@@ -24,12 +24,12 @@ import numpy as np
 
 from nemo.collections.asr.metrics.wer import word_error_rate
 from nemo.collections.asr.models import ClusteringDiarizer
-from nemo.collections.asr.parts.utils.manifest_utils import write_rttm2manifest
 from nemo.collections.asr.parts.utils.speaker_utils import (
     audio_rttm_map,
     get_uniqname_from_filepath,
     labels_to_rttmfile,
     rttm_to_labels,
+    write_rttm2manifest,
 )
 from nemo.utils import logging
